Update duplicate code
